app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.1.3"
black = "^22.8.0"
isort = "^5.10.1"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.21.0"
flake8 = "^5.0.4"
pytest-black = "^0.3.12"
pytest-cov = "^4.0.0"
coloredlogs = "^15.0.1"
mypy = "^0.981"
requests = "^2.28.1"
python-decouple = "^3.6"
Faker = "^15.0.0"
SQLAlchemy = "^1.4.41"
ipykernel = "^6.16.0"
databases = "^0.6.1"
uvicorn = "^0.20.0"
fastapi = "^0.88.0"
redis = "^4.4.0"
async-sqlalchemy = "^1.0.0"
alembic = "^1.9.0"
python-multipart = "^0.0.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
